+++ utils/maputils/qbsp.c
@@ -131,10 +131,10 @@ CopyWinding
 */
 winding_t *CopyWinding (winding_t *w)
 {
-	int			size;
+	size_t			size;
 	winding_t	*c;
 
-	size = (int)((winding_t *)0)->points[w->numpoints];
+	size = (size_t)((winding_t *)0)->points[w->numpoints];
 	c = malloc (size);
 	memcpy (c, w, size);
 	return c;
@@ -401,7 +401,7 @@ NewWinding
 winding_t *NewWinding (int points)
 {
 	winding_t	*w;
-	int			size;
+	size_t			size;
 
 	if (points > MAX_POINTS_ON_WINDING)
 		Error ("NewWinding: %i points", points);
@@ -410,7 +410,7 @@ winding_t *NewWinding (int points)
 //	if (c_activewindings > c_peakwindings)
 //		c_peakwindings = c_activewindings;
 
-	size = (int)((winding_t *)0)->points[points];
+	size = (size_t)((winding_t *)0)->points[points];
 	w = malloc (size);
 	memset (w, 0, size);
