+++ utils/maputils/vis.c
@@ -189,12 +189,12 @@ void FreeWinding (winding_t *w)
 winding_t *NewWinding (int points)
 {
 	winding_t	*w;
-	int			size;
+	size_t			size;
 
 	if (points > MAX_POINTS_ON_WINDING)
 		Error ("NewWinding: %i points", points);
 
-	size = (int)((winding_t *)0)->points[points];
+	size = (size_t)((winding_t *)0)->points[points];
 	w = malloc (size);
 	memset (w, 0, size);
 
@@ -241,10 +241,10 @@ CopyWinding
 */
 winding_t *CopyWinding (winding_t *w)
 {
-	int			size;
+	size_t			size;
 	winding_t	*c;
 
-	size = (int)((winding_t *)0)->points[w->numpoints];
+	size = (size_t)((winding_t *)0)->points[w->numpoints];
 
 //	c = malloc (size);
 	c = NewWinding(w->numpoints);
